refactor(docker): split client implementation by concern
